// src/app/news/models/vaccines.model.ts
export interface Vaccines {
  administered: number;
  people_vaccinated: number;
  country: string;
  population: number;
  sq_km_area: number;
  life_expectancy: number;
  elevation_in_meters: number;
  continent: string;
  abbreviation: string;
  location: string;
  iso: number;
  capital_city: string;
  lat: number;
  long: number;
  updated: string;
}